Revert "Fix warnings in xrGame"
This reverts commit 5449a97fd30cb1237d7501d8fedf222e6ab8eba3.

+++ src/xr_3da/xrGame/CarWeapon.cpp
@@ -196,7 +196,7 @@ float CCarWeapon::FireDirDiff()
 	Fvector d1,d2;
 	d1.set(m_cur_x_rot,m_cur_y_rot,0).normalize_safe();
 	d2.set(m_tgt_x_rot,m_tgt_y_rot,0).normalize_safe();
-	return (float) rad2deg( acos(d1.dotproduct(d2)) );
+	return rad2deg( acos(d1.dotproduct(d2)) );
 }
 
 const Fvector&	CCarWeapon::get_CurrentFirePoint()
