+++ src/xr_3da/xrGame/PHAICharacter.cpp
@@ -46,7 +46,7 @@ bool CPHAICharacter::TryPosition(Fvector pos,bool exact_state){
 	float	rest = 0.f;
 
 	float	parts = disp_mag / disp_pstep ;
-	fsteps_num = (float) floor ( parts );
+	fsteps_num = floor ( parts );
 	steps_num = iFloor ( parts );
 	if( steps_num > max_steps )
 	{
